V5.8
V5.8 Added support for ResearchBodies mod (MUST HAVE ResearchBodies V1.6
or above).
If ResearchBodies mod is installed then:-
Telescope and ChemCam contracts will NOT generate for bodies that have
yet to be discovered.
If you take a picture with a Telescope of a body that is yet to be
discovered AND transmit/recovery it, it will become discovered.
Subsequent pictures taken and transmitted/recovered of a already
discovered body will add 20 research to that body.
The TST Telescopes act the same as the ResearchBodies telescope, IE:they
have the Research button in their right click menu and will discover
bodies they are pointed at in the same way as the ResearchBodies
telescope does.
The TST GUI camera-window "Show Bodies" list will NOT show un-discovered
bodies or galaxies.